chore_: move localbackup code to wallet service
